Support 16KB pages for apps targeting Android 15+ (#53)
* Update serious_python dependency ref in pubspec.yaml

Changed the git ref for the serious_python dependency from 'main' to 'android-16k' to use a different branch or commit.

* Update webview_flutter_android to version 4.10.1

Bumped the webview_flutter_android dependency from ^4.0.0 to 4.10.1 in pubspec.yaml to ensure compatibility with the latest features and bug fixes.

* Add conditional override for webview_flutter_android

Introduced a conditional block in pubspec.yaml to override the webview_flutter_android dependency only if 'flet_webview' is present in the dependencies. This improves template flexibility for different dependency configurations.

* Update webview_flutter_android dependency handling

Moved webview_flutter_android to be conditionally included based on the presence of 'flet_webview' in dependencies. Cleaned up commented-out geolocator dependencies.